pyproject.toml
[project]
name = "movie-anime-discovery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.2",
    "django-tailwind[cookiecutter,honcho,reload]>=4.4.2",
    "httpx>=0.28.1",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "django-stubs>=6.0.2",
    "pytest>=9.0.2",
    "ruff>=0.15.8",
    "ty>=0.0.26",
]
